PDA-60 separating KieTask from KieTaskDetails in order to match Task List Columns endpoint response with actual task list DTO;
